Refactor WTA helper runtime ownership
Move the per-pane helper TUI runtime out of main.rs and into the helper module without changing behavior. Document the broader incremental architecture refactor and correct the crate description.


Copilot-Session: b3f5f1bc-78dd-4424-9417-88029e6535ad
