+++ tools/wta/Cargo.toml
@@ -2,7 +2,7 @@
 name = "wta"
 version = "0.1.0"
 edition = "2021"
-description = "Windows Terminal Agent — ACP TUI client / MCP tool server"
+description = "Windows Terminal ACP agent bridge and control CLI"
 build = "build.rs"
 
 [[bin]]
