+++ tools/wta/src/helper/mod.rs
@@ -8,18 +8,18 @@
 // is passed in and speaks ACP JSON-RPC over it. From the helper's
 // perspective, master IS the agent.
 //
-// All App / event-loop machinery is reused from `run_default_tui_over_pipe`;
-// the only delta is the ACP transport, which is selected via
-// `--connect-master` and threaded down through `run_acp_app`, which drives
-// `run_acp_client_over_pipe` against the wta-master pipe.
+// The helper implementation lives in `helper/tui.rs`: WT connection,
+// terminal lifecycle, channel wiring, and the event loop.
 
 use anyhow::Result;
 
 use crate::Cli;
 
+mod tui;
+
 /// Helper-mode entry point. Routes the ACP traffic through a named pipe
 /// to the wta-master singleton instead of spawning a private agent CLI
 /// subprocess.
 pub async fn run_helper_mode(cli: Cli, pipe_name: String) -> Result<()> {
-    crate::run_default_tui_over_pipe(cli, pipe_name).await
+    tui::run_default_tui_over_pipe(cli, pipe_name).await
 }
